+++ src/main/java/org/example/repository/GemstoneRepository.java
@@ -8,12 +8,12 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GemstoneRepository extends JpaRepository<Gemstone, Integer> {
+public interface GemstoneRepository extends JpaRepository<Gemstone, Long> {
     @Query(value = "UPDATE is_delete = true FROM gemstone g WHERE g.id = ?1",nativeQuery = true)
-    void softDeleteById(int id);
+    void softDeleteById(long id);
     @Query(value = "SELECT * FROM gemstone g WHERE g.is_deleted = true",nativeQuery = true)
     List<Gemstone> findAllDeletedGemstones();
 
     @Query(value = "UPDATE is_delete = false * FROM gemstone g WHERE g.id = ?1",nativeQuery = true)
-    void restoreDeletedGemstone(int id);
+    void restoreDeletedGemstone(long id);
 }
